backend/requirements.txt
strands-agents
strands-agents-tools
boto3
neo4j
fastapi
uvicorn
python-dotenv
ddtrace
python-multipart
requests
